+++ apps/admin/src/components/lab/lab-table-view.tsx
