+++ apps/admin/src/components/professor/professor-table-view.tsx
